+++ MenuState.cs
@@ -23,29 +23,29 @@ namespace RaahnSimulation
 	    {
 	        base.Init(sim);
 
-	        float charWidth = (float)context.GetWindowWidth() * Utils.CHAR_WIDTH_PERCENTAGE;
-	        float charHeight = (float)context.GetWindowHeight() * Utils.CHAR_HEIGHT_PERCENTAGE;
+	        double charWidth = (double)context.GetWindowWidth() * Utils.CHAR_WIDTH_PERCENTAGE;
+	        double charHeight = (double)context.GetWindowHeight() * Utils.CHAR_HEIGHT_PERCENTAGE;
 
 	        title = new Text(context, Utils.WINDOW_TITLE);
 	        title.SetWindowAsDrawingVec(true);
-	        title.SetCharBounds((float)context.GetWindowWidth() / 2.0f, (float)context.GetWindowHeight() - charHeight, charWidth, charHeight, true);
+	        title.SetCharBounds((double)context.GetWindowWidth() / 2.0f, (double)context.GetWindowHeight() - charHeight, charWidth, charHeight, true);
 
             version = new Text(context, Utils.VERSION_STRING);
             version.SetWindowAsDrawingVec(true);
             version.SetCharBounds(0.0f, 0.0f, charWidth, charHeight, false);
 
-            float startSimWidth = charWidth * Utils.START_SIM.Length;
+            double startSimWidth = charWidth * Utils.START_SIM.Length;
 
 	        startSim = new Button(context, Utils.START_SIM);
 	        startSim.SetWindowAsDrawingVec(true);
-	        startSim.SetBounds((float)context.GetWindowWidth() / 2.0f, title.windowPos.y - (2.0f * charHeight), startSimWidth, charHeight, true);
+	        startSim.SetBounds((double)context.GetWindowWidth() / 2.0f, title.windowPos.y - (2.0f * charHeight), startSimWidth, charHeight, true);
             startSim.SetOnClickListener(startSimOnClick);
 
-            float startMapWidth = charWidth * Utils.START_MAP.Length;
+            double startMapWidth = charWidth * Utils.START_MAP.Length;
 
 	        startMap = new Button(context, Utils.START_MAP);
 	        startMap.SetWindowAsDrawingVec(true);
-	        startMap.SetBounds((float)context.GetWindowWidth() / 2.0f, startSim.windowPos.y - (2.0f * charHeight), startMapWidth, charHeight, true);
+	        startMap.SetBounds((double)context.GetWindowWidth() / 2.0f, startSim.windowPos.y - (2.0f * charHeight), startMapWidth, charHeight, true);
 	        startMap.SetOnClickListener(startMapOnClick);
 
 	        AddEntity(title, 0);
